+++ src/plugin.ts
@@ -1,3 +1,4 @@
+import { Linter } from "eslint";
 import * as parserJsonc from "jsonc-eslint-parser";
 import { createRequire } from "node:module";
 
@@ -11,13 +12,15 @@ import { rules as requireRules } from "./rules/require-properties.js";
 import { rule as restrictDependencyRanges } from "./rules/restrict-dependency-ranges.js";
 import { rule as sortCollections } from "./rules/sort-collections.js";
 import { rule as uniqueDependencies } from "./rules/unique-dependencies.js";
+import { rule as validAuthor } from "./rules/valid-author.js";
+import { rule as validBin } from "./rules/valid-bin.js";
 import { rule as validLocalDependency } from "./rules/valid-local-dependency.js";
 import { rule as validName } from "./rules/valid-name.js";
 import { rule as validPackageDefinition } from "./rules/valid-package-definition.js";
 import { rule as validRepositoryDirectory } from "./rules/valid-repository-directory.js";
 import { rule as validVersion } from "./rules/valid-version.js";
 
-const require = createRequire(import.meta.url || __filename);
+const require = createRequire(import.meta.url);
 
 const { name, version } = require("../package.json") as {
 	name: string;
@@ -33,6 +36,8 @@ const rules: Record<string, PackageJsonRuleModule> = {
 	"restrict-dependency-ranges": restrictDependencyRanges,
 	"sort-collections": sortCollections,
 	"unique-dependencies": uniqueDependencies,
+	"valid-author": validAuthor,
+	"valid-bin": validBin,
 	"valid-local-dependency": validLocalDependency,
 	"valid-name": validName,
 	"valid-package-definition": validPackageDefinition,
@@ -54,11 +59,34 @@ const rules: Record<string, PackageJsonRuleModule> = {
 	},
 };
 
-const recommendedRules = Object.fromEntries(
-	Object.entries(rules)
-		.filter(([, rule]) => rule.meta.docs?.recommended)
-		.map(([name]) => ["package-json/" + name, "error" as const]),
-);
+const baseRecommendedRules = {
+	...Object.fromEntries(
+		Object.entries(rules)
+			.filter(([, rule]) => rule.meta.docs?.recommended)
+			.map(([name]) => ["package-json/" + name, "error" as const]),
+	),
+} satisfies Linter.RulesRecord;
+
+const recommendedRules = {
+	...baseRecommendedRules,
+	// As we add more `valid-*` rules, we should prevent this legacy rule from
+	// also reporting the same errors.
+	"package-json/valid-package-definition": [
+		"error",
+		{
+			// Create a list of properties to ignore based on the valid-* rules
+			// we currently have. Once we've fully covered what `valid-package-definition`
+			// checks, we can remove it from the `recommended` config entirely.
+			ignoreProperties: Object.entries(baseRecommendedRules)
+				.filter(
+					([name]) =>
+						name.startsWith("package-json/valid-") &&
+						name !== "package-json/valid-package-definition",
+				)
+				.map(([name]) => name.replace("package-json/valid-", "")),
+		},
+	],
+} satisfies Linter.RulesRecord;
 
 export const plugin = {
 	configs: {