+++ vitest.config.ts
@@ -5,9 +5,9 @@ export default defineConfig({
 		clearMocks: true,
 		coverage: {
 			all: true,
-			exclude: ["lib", "src/index.ts", "src/rules/index.ts"],
+			exclude: ["lib", "src/index.ts", "src/rules/index.ts", "src/tests"],
 			include: ["src"],
-			reporter: ["html", "lcov"],
+			reporter: ["html", "lcov", "text"],
 		},
 		exclude: ["lib", "node_modules"],
 		setupFiles: ["console-fail-test/setup"],